+++ src/routers/home.ts
@@ -6,6 +6,7 @@ import Mangasee from "../scrapers/mangasee";
 import { setMangaProgress } from "../util/getMangaProgress";
 import getReading from "../util/getReading";
 import db from "../db";
+import { getLists } from "../util/lists";
 
 router.get("/", async (req, res) => {
 
@@ -20,6 +21,7 @@ router.get("/", async (req, res) => {
 	// Set progress for popular manga
 	await Promise.all(popular.map(setMangaProgress));
 
+	// Get reading
 	let reading = await getReading();
 
 	res.render("home", {
