+++ views/partials/list.handlebars
@@ -1,0 +1,15 @@
+{{#if list.entries}}
+	<details open id="{{list.slug}}" class="section lists big-summary decreased-margin">
+		<summary>
+			<div class="summary-inner">
+				<h2 class="sub-title">{{list.name}}</h2>
+				<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-chevron-down"> <polyline points="6 9 12 15 18 9"></polyline> </svg>
+			</div>
+		</summary>
+		<div class="manga-grid">
+			{{#each list.entries}}
+			{{>manga-card manga=this.data}}
+			{{/each}}
+		</div>
+	</details>
+{{/if}}
